+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ## Unreleased
 
-- TASK-004 - Refactor module namespace
 
 ## v1.0.0 - 2022-02-22
 
 - TASK-001 - Create composer needed files
 - TASK-002 - I'm going to try this using an after plugin interceptor for the core method Magento\Sales\Model\Order\Pdf\Invoice::getPdf()
 - TASK-003 - Create CMS Block static content through data patch
+- TASK-004 - Refactor module namespace
+- TASK-005 - Data helper to generate the PDF block layout for the invoice and Plugin to apply the block